se agregan tareas a procesos de un proyecto

+++ app/Http/Controllers/ActivityController.php
@@ -1,30 +1,39 @@
 <?php namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
-use App\Models\Activity;
+use App\Models\Activities;
 use Input;
+use Response;
 
 class ActivityController extends BaseController {
 
     public function load() {
         $id = Input::get('id');
         if($id == 0){
-            return json_encode(Activity::all());
+            return json_encode(Activities::all());
         }
-        return json_encode(Activity::find($id));
+        return json_encode(Activities::find($id));
     }
+
     public function create() {
-        return json_encode(parent::baseCreate(new Activity));
+        $activity = new Activities();
+        $activity->name = Input::get('name');
+        $activity->process_id = Input::get('process_id');
+        $activity->save();
+        return Response::json($activity);
     }
 
     public function update($id) {
-        $activity = Activity::find($id);
+        $activity = Activities::find($id);
         return json_encode(parent::baseUpdate($activity));
     }
 
-    public function destroy($id) {
-        $activity = Activity::find($id);
-        $activity->delete();
+     public function destroy() {
+        $id = Input::get('id');
+        $activity = Activities::find($id);
+        if(is_object($activity)){
+            $activity->delete();    
+        }
         return json_encode($id);
     }
